// src/store/mutations.type.ts
export enum MUTATIONS {
    SET_QUIZ = "setQuiz",
    ADD_THEME = "addTheme",
    DELETE_THEME = "deleteTheme",
    ADD_QUESTION = "addQuestion",
    DELETE_QUESTION = "deletQuestion",
    UPDATE_THEME_LABEL = "updateThemeLabel",
    UPDATE_QUESTION_TEXT = "updateQuestionText",
    UPDATE_QUESTION_ANSWER = "updateQuestionAnswer",
    UPDATE_QUESTION_MAX = "updateQuestionMax",
    UPDATE_QUESTION_INCREMENT = "updateQuestionIncrement"
};